[ci] Only run FPGA tests on pull requests.
This is to minimize the load on the CI infrastructure. Post-submit
testing should be time bound to no more than a few  minutes on resources
with very limited availability. This change takes the drastic measure of
not running any FPGA tests, but a reduced test suite for post submit
can be added in future (e.g. a `test_suite` with optimized coverage).

Signed-off-by: Miguel Osorio <miguelosorio@google.com>