[crypto] Move P-384 point-validity checks into OTBN.
Now that we have the IMEM space for it, it makes sense to push point-validity
checks closer to the use of the point in ECDSA verify and ECDH. This removes a
lot of places where the validity check and point use are decoupled, which could
lead to security holes. It also simplifies the code quite a bit and saves a
chunk of code size.

Signed-off-by: Jade Philipoom <jadep@zerorisc.com>
(cherry picked from commit 810945a2cac4af6e7368e1013295570cc6da6d20)