[sival] Re-enable passing tests marked broken in SiVal ROM EXT
Several tests were previously marked as broken in the SiVal ROM EXT
environment due to failures that have since been fixed, meaning that the
tests are now consistently passing in nightlies (for around a month),
but they are still marked as broken.

As these tests appear to be passing fine, this commit removes the broken
tag.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>