[fpga] Add ability to issue POR to self via SW
Co-opt the pad attributes in sensor_ctrl for FLASH_TEST_MODE0 to
enable firmware on the FPGA to reset itself with a POR. The FPGA does
not have the FLASH_TEST_MODE0 pad, and the ability to give itself a
POR allows us to add firmware to fix up inconsistent starting points
for the bitstream vs. a real provisioned chip.

This change is motivated by a need to fix up the flash info pages to
reflect the expected chip state when in PROD. It is possible to splice
VMEM images as is currently done for OTP and ROM, but calls to
updatemem appear to consume an undesirable amount of time.

Signed-off-by: Alexander Williams <awill@opentitan.org>