[sival,bazel] Add SiVal ROM EXT exec env for PMOD tests
The current SiVAL ROM EXT execution environment uses the hyperdebug
(hyper310) interface and bitstream, which does not allow all the pins
that are needed for the PMOD tests (e.g. the SPI Macronix flash test) to
be used.

This commit adds a new execution environment that is private to the PMOD
Bazel BUILD file that extends the existing `fpga_cw310_sival_rom_ext`
environment to use the universal cw310 bitstream and interface, such
that this test can be run with the SiVal changes (e.g. OTP) and in ROM
EXT.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>