[dv/dpi] jtagdpi plusarg for initial SRST value
The jtagdpi will check for a non-zero value in the plusarg
"+jtagdpi_assert_srst=1" to determine if the functional reset wire
should be asserted initially. This makes it possible for a client
connection to be established without racing against the simulation.

The default behavior (when the plusarg is not present or its value is
0) is still for the active low functional reset output from jtagdpi to
be deasserted initially.

Signed-off-by: Kip Walker <kip@rivosinc.com>