[dv/dpi] tcp_server error out on port conflict
When the tcp_server is unable to create its listening socket, the
simulation now exits. Previously, an error would be printed but the
simulation would continue. Clients like dmidpi and jtagdpi should only
call tcp_server_create if there's a chance a client will connect, and
if the listening port number is being selected to avoid conflict with
other processes on the same machine.

Signed-off-by: Kip Walker <kip@rivosinc.com>