[rules] Add support for templates in linker files
Linker files are not preprocessed, which makes it hard to support
several tops. In particular, including different files based on
the top, or using a different region is very difficult.

The `ld_library` rule is modified to take a new substitution
dictionary as input which is subject to location expansion. This
allows for example to do in the linker script:
```
INCLUDE @@MYFILE@@
```
and then in the BUILD file:
```
ld_library(
  ...,
  substitutions = {
    "@@MYFILE@@": "$(location //path/to/my/file)",
  },
)
```

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>