[HACK] Make sim_dv exec_env compatible with Darjeeling
Due to a limitation of the design of exec_env, it is not possible
to create two `sim_dv` environments and compile binaries for both
(this is because they will share the same provider). Instead,
create a darjeeling exec_env for sim_dv, and a new alias in
`//hw/top`. Then replace the earlgre-specific exec_env
in `EARLGREY_TEST_ENVS` by this new one.

This means that with this change, any previously sim_dv-enabled
test can be compile on darjeeling simply by passing
--//hw/top=darjeeling on the command line.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>