[HACK] Always transition to earlgrey in golden toolchain
Unfortunately, opentitantool is far from compiling with darjeeling
due to many explicit dependencies on earlgrey. At the same time,
opentitantool provides many operations which are not earlgrey-specific
at all such as flash image generation, signing, modid checks, etc

This commit introduces a necessary hack: in the host transition
for the golden toolchain, always set the top to earlgrey. This way
we can ensure that opentitantool will compile. For now this does
not introduce any problem but this is not a proper fix.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>