[util,tlgen] Make --outdir a required argument and exit if empty
Before this change, you get a confusing error message if you don't
supply it. The original author also obviously imagined that
log.error() would end the program, but it doesn't! Bail properly if
needed.

Finally, tighten up the test with is_dir() because Path('') gives the
current directory (which is a directory...)

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>