[rv_dm,sival] Fix `rv_dm_mem_access_delayed_enable`
The `rv_dm_mem_access` test was not working in the
`dev_rv_dm_delayed_enable` test configuration. This is because the test
was not written to handle the additional logic required for this
configuration.

To fix this, before the host side attempts to connect via JTAG, it waits
for a message ("DEBUG_MODE_ENABLED") from the device, which will
configure the RV DM to enable the late debug feature (debug mode). This
should only be available using the OTP image with the late_debug feature
enabled, in the `delayed_enable` test configuration.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>