[bitstream] Collect MMI data into one file
Reduce BUILD file boilerplate by placing all memory map info in a single
file. Use a separate Processor element for each map, so the InstData
attribute can be used as a key to identify the memory to be updated.
This construction also permits multiple memories to be spliced in a
single call to updatemem, though that is not implemented in this commit.

Adjust MMI dump to use a regex for mem type and a dictionary for each
memory's parameters. For the mem type, some memories in the FPGA can
consist of both RAMB36 and RAMB18 types, depending on the width of the
array. The regex allows more flexible memory cell detection.

Rev up the bitstream manifest schema to v3 so the memories contained in
the MMI file are described. Add a function to collect MMI data from v2
cache entries and rewrite in the expected v3 format, with the MMI data
in a single file.

Finally, drop support for bitstream cache entries that do not contain a
manifest. The tools to create a manifest have been in the repo for
awhile, and as of this commit's creation, no such cache entries remain
in the public cloud storage bucket.

Signed-off-by: Alexander Williams <awill@opentitan.org>
(cherry picked from commit 9e1ef80bca3d6648b1091efd1e5e7b3776a220f2)