[fpga] Dump flash info maps to the MMI file
For earlgrey, export the flash info memory maps to the MMI file. Add
knowledge of these maps existing to the bitstream cache entries. The
flash info arrays are identified by a key in this format:

    flash<FlashBank>_info<InfoType>

Add KEEP_HIERARCHY to the prim_ram_1p instances representing the flash
info arrays. This helps keep intelligible hierarchical paths to the
memories, so we can readily select the correct cells.

This lays groundwork for a future PR where we can splice these memories.

Signed-off-by: Alexander Williams <awill@opentitan.org>