[prim,rtl] Rephrase a term in prim_reg_cdc.sv
If DstWrReq is false then the src_update signal cannot be true. This
is because it is driven by the src_update_o port on a prim_reg_cdc_arb
instance. If DstWrReq is false then the instance wires src_update_o to
zero (in the gen_passthru generate block).

This rewrite will behave equivalently in the same way as before but
will no longer generate any conditional coverage terms that depend on
src_update if DstWrReq is false.

Note that if DstWrReq is zero then the src_update and busy signals are
constant zero and are no longer actually used. To avoid a lint error
from the unused signals in this situation, they get "sampled" in the
gen_passthru block.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>