[topgen] Add class to load a complete topcfg properly
At the moment, topgen simplies serialized the complete topcfg using
hjson.dumps() but this poses a few problems. The biggest one is that
it is a mix of manual conversion to dict (when calling as_dict()
in the various topgen functions) and automatic (when calling _asdict()
from hjson.dumps). Furthermore, it turns out that we are missing some
fields that probably were never added to _asdict().

This commit introduces a new class (CompleteTopCfg) whose sole purpose
is to take the produced Hjson and reconstruct an in-memory topcfg that
is *exactly* equivalent (in Pythonic types) to the ones that was dumped.
This requires to sometimes reconstruct some classes, sometimes not.
Classes that need to be reconstruct get a new method (fromdict) and
the CompleteTopCfg does as much automatic deserializing as possible,
then some manual fixing.

There is also a small tweak to how clock groups are handled: in
alert_lpgs, the clock groups are all elements of the clocks.groups
but because of the naive serializing, a full copy is made. This
commit introduces GroupProxy that behaves like a Group for all
purposes but serializes only to the group name. This makes the
resulting hjson clearer and the deserialization easier.

Since this process is quite fragile, when topgen is running it will
actually check that this works by dumping the Hjson, reloading it
with the CompleteTopCfg and then checking that the two are equivalent.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>