[ipgen,otp_ctrl] Turn a few more files into templates and fix paths
The template files are almost identical to the previous ones, except
- they use ${topname} instead of a specific name
- consecutive '#' are escaped so they are not considered hugo comments;
  for example, ### is replaced by ${"###"}
- text bracketed by CMDGEN is removed

Adjust some relative paths to account for the change in the file locations.

Use {self_dir} to simplify some paths: it points to the location of the hjson
file used to process the contents of a given file.

Fix some incorrect links to docs.

Part of #25019

Signed-off-by: Guillermo Maturana <maturana@opentitan.org>