[autogen,tests] Rework plic_all_irqs_test setup
The setup is very earlgrey specific, assuming exactly 23 IRQ
peripheral. This commit changes the setup so that each test will
test 10 peripherals, except for the last test that will capture all
remaining. This ensures that we always test every peripheral, and
if the last test becomes too large than we just need to bump
NR_IRQ_PERIPH_TESTS.

This changes the name of the tests which are now numbered 0, 1, etc
instead of 0, 10, 20 which made little sense anyway.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>