[pwm] Eliminate inter-channel crosstalk
Prevent changes to the 'pwm_en' and 'invert' registers impacting
channels that are already enabled and running.

Introduce a register stage on each of the pwm output signals
to prevent glitches being produced by the fairly involved
combinational logic at the channel outputs. This delays each
output signal by a single core clock cycle but the phases of
the output signals are unaffected.

Signed-off-by: Adrian Lees <a.lees@lowrisc.org>