[pwm] Corrections to heartbeat and blinking modes
Switch over to a new duty cycle only when not producing a pulse;
the original logic did not consider the effect of 'phase_delay.'

We switch to the new duty cycle when the phase counter is just
about to cross the 'phase_delay' threshold. For this we need to
know the next phase that we will see, but conveniently that is
available within the parent module for all channels to use.

Blink mode is less expensively implemented using a single down
counter that toggles between 'x' and 'y' values rather than using
addition and comparison. For this we need just remember whether
blinking is in its 'y' phase or not.

Eliminate the redundancy of additional counting for heartbeat mode,
and instead use the regular 'x' phase down counting. Heartbeat
blinking remains always in the 'x' phase.

Since we already have to compare 'duty_cycle_a_i' and
'duty_cycle_b_i' to ascertain which is higher, we can simplify
the comparison against upper and lower duty cycles in heartbeat
mode by determining the '_max' and '_min' values.

Signed-off-by: Adrian Lees <a.lees@lowrisc.org>