[sram_ctrl,dv] Fix tl_errors test
With lowRISC/opentitan#24830 the tl_errors test was adapted to also
include accesses to CSRs. As the sram_ctrl RTL returns different
d_data (either '1 or '0) when targeting the SRAM or the registers,
the tl_errors test was failing since then.

This commit fixes the test by adapting the expected data depending
on whether it is a CSR or SRAM access.

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>