[rules] Creates new rules to extract information from top description
Now that the top description is available from the opentitan_top macro,
we can create rules to extract it in a form that is more convenient for
the build system. This commit introduces three rules:
- two to extract the top's C library and linker files
- one to construct a top description in the form of a provider

This rules essentially reformats the opentitan_top's struct content
as a provider. The reason for this indirection is that for rules
that will depend on the top description in the analysis phase (and
not the loading phase), getting the information via a provider is
much cleaner and useful that via a struct.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>