[rom_ctrl, rtl] tlul_adapter_sram case 00 for tl_i_int and error_internal
It is impossible to see !tl_i_int.a_valid & !error_internal. When a_valid is false,
it is seen in u_err inside tlul_adapter_sram.sv. This will make a_config_allowed
inside u_err false as it depends on addr_sz_chk, mask_chk and fulldata_chk.
All of them are false if a_valid is false. This means that the first term inside
err_o in u_err is false which raised err_o.
Next, err_o would become visible to error_det as tlul_error inside tlul_adapter_sram
and it will be true if tlul_error is true.
error_det will then be seen as error_i to u_sram_byte inside tlul_adapter_sram. It
is the part of the definition for error_o in u_sram_byte. Unless stall_host is true,
error_i comes out as error_o. Then error_o would be seen as error_internal to
tlul_adapter_sram and the conditional statement is not able to cover 00 case for this
reason.

This fix remove the item (tl_i_int.a_valid) from the statement to subtract the coverage
hole and then adds an assertion to make sure the 00 never happens.

Signed-off-by: Kinza Qamar <kqzaman@lowrisc.org>