[rom_ctrl,dv] Fix how we splat fifo contents in common vseq
This code was added by cc0abbfe3e3 and I think it was assuming that we
only ever run this after reset (when the FIFO will be empty). It turns
out that this isn't quite true and there is a window where we can
write rubbish to the FIFO when it isn't empty, causing a sporadic
failure in rom_ctrl_stress_all_with_rand_reset.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>