[bazel] Temporarily ignore old vendored crate files
These are no longer used, but are left in the repository to avoid an
unreviewable pull request. They will be removed in a later commit. In
the mean time, Bazel is asked to ignore them so that it can build the
rest of the graph. This also ensures they aren't used / accessible to
the rest of the Bazel project.

Signed-off-by: James Wainwright <james.wainwright@lowrisc.org>