[crypto] Harden RSA modexp subroutine.
Add some SCA hardening measures to the modexp subroutine that selects either
the squared or squared-and-multiplied result in square-and-multiply-always
modular exponentiation.

Although the rest of the code is not hardened, the hardening for RSA will
likely preserve this subroutine, so hardening it is still useful.

Signed-off-by: Jade Philipoom <jadep@zerorisc.com>
(cherry picked from commit 8c638309857e08ef5b4002d91315d4971f1f0872)