[dv] Avoid timing out if we call tl_access_sub lots in parallel
Before this change: when we start a sequence to perform a CSR access,
we wrapped it in a DV_SPINWAIT with a timeout of e.g.
tl_access_timeout_ns. This is a sensible idea! If the TL bus dies,
we'd like the test to time out and fail.

But we might call e.g. csr_rd many times in parallel. If we do so,
we'll end up enqueuing many copies of tl_seq. These each take some
time to run and a very wide queue will cause the DV_SPINWAIT above to
time out. That's a bit silly. There's no problem with the hardware:
we're just doing a load of stuff, which takes a long time.

To avoid this problem, we change the timeout so that it only starts
when the sequence in question has actually been granted a slot by the
sequencer.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>