[rtl, aon_timer] Reset core.prescale_count on wkup_ctrl write
The AON timer core has an internal counter `prescale_count_q` which is
not reset upon each write to wkup_ctrl.
This causes issues if the WKUP timer is re-configured to a lower
prescaler value than it was before, and the `prescale_count_q` variable
is greater than the new prescaler value, forcing the
`prescale_count_q` counter to overflow itself causing a non-spec
compliant behavior.

In order to mitigate the above, the internal `prescale_count_q` is
reset after wkup_ctrl write. Thus, the aon timer is forced to count
from scratch the new intended prescaler value.

Signed-off-by: Antonio Martinez Zambrana <amz@lowrisc.org>