[util] Fix audit_sec_mmio_calls to use the new IP header rules
With the new rule, a very easy way to get the list of all Hjson
files for IPs is to look at the `ip_hjson` attribute of the
definition of //hw/top:top_<name>_desc. Since this script kind
of assumes earlgrey anyway, explicitely point to the earlgrey
description of now.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>