[ci] Move CW305 bitstream build to machine without license
The CW305 uses an Artix-7 FPGA which does not require an enterprise
Vivado license to synthesise for. We limit the number of machines on
the `-bitstream` label to match our number of licenses, however CW305s
are taking up these spots unnecessarily.

This commit moves them to the `-vivado` machines which have access to
Vivado Standard edition only.

Signed-off-by: James Wainwright <james.wainwright@lowrisc.org>