[rom_ext] Enable Oz to reduce firmware size
This change enables Oz for rom_ext to reduce code size (43084B -> 41868B).

Since some targets can't be built with LLD, this commit introduces a new
attribute, `with_features`, to the `opentitan_binary` rule to allow
specifying additional features for the target.

LTO is also ready to be enabled for those targets, but we found no size
improvement (41868B v.s. 41896B), so we keep it off at the moment.

Change-Id: I5a8e7e7fcbb7d1e435b987d955a3e5ff8e4b7383
Signed-off-by: Yi-Hsuan Deng <yhdeng@google.com>