[bazel] Introduce an extension for python in repository rules
Currently, it is quite tricky to use the python interpreter in
the repository rules (like in bitstream_workspace.py) because
the repositories created by rules_python are not really meant
to be used in the loading phase. With blzmod, this has become
quite cumbersome and so this commit introduces a clean
mechanism to do so.

The first is to create an extension. This extension has a single
method to register a python interpret and some packages which
are import by pip. The extension will then create a repository
containing a single executable which acts as python interpreter
taking care of all the detais.

Strictly speaking, this could have been achieved without an
extension but it would then require to manipulate the pythonpath
at every call site and create the required labels every time.
By putting everything into an extension, we get to do it once
and make it very easy to use everywhere.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>