[bazel] Capture output of updatemem to make it less verbose
updatemem likes to print a lot of things which makes the CI output
less readable for no particular benefit. This commit introduces a
bash script that runs updatmem and captures the output. If an error
occurs, it will print the path to the output and error logs. Those
files are also added to the output groups of the bitstream_splice
rules so that they can be queried if needed.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>