SQUASH ME! [topgen] Partially fix up alert / irq domains
Fix up alert and irq domains to enable generating collateral where the
alerts and irqs don't necessarily generate from devices in the same
address space as the CPU. Darjeeling has bus bridges that make servicing
those interrupts possible (indirect reachability to the IP).

The autogenerated tests will surely be bonkers, but that is to be fixed
up later.

Signed-off-by: Alexander Williams <awill@opentitan.org>