[dv,pwm] Resurrect pwm_alert_test
Switching the alert agents to synchronous operation resulted in
the `pwm_alert_test` failing because the alert_sender modules in
the PWM have - necessarily - an asynchronous connection.

Instead we now suspend the alert agents whilst the TL-UL clock
is disabled so that the two ends of the differential buses do
not get out of step. The alert agents are thus restored to
asynchronous operation.

Signed-off-by: Adrian Lees <a.lees@lowrisc.org>