[hmac, rtl] Increase coverage through minor RTL restructure
In light of the V3 signoff, this commits adapts certain RTL parts that
induced uncoverable coverage holes. There are two types of changes:

  1. Removal of dangling else branches that are impossible to cover.
  2. Removal of uncoverable FSM default state assignments.

Both changes have no functional impact and are only cosmetic in nature
in order to attain the demanded coverage thresholds. See #24692, for a
detailed discussion of these changes.

Signed-off-by: Andrea Caforio <andrea.caforio@lowrisc.org>