[crypto] Fix the KMAC-KDF implementation to return the output in shares.
Previously, the KMAC driver always unmasked the shares of the digest returned
by the KMAC block. This is OK for hash functions like SHA-3, but problematic
for KMAC-KDF where the output is key material. The implementation also did not
modify the second half of the keyblob buffer in KMAC-KDF, which would cause
correctness issues if those values were not initialized to zero.

This commit refactors the KMAC driver to allow returning the digest in masked
form, and adjusts KMAC-KDF to use this new capability.

Signed-off-by: Jade Philipoom <jadep@zerorisc.com>
(cherry picked from commit c676ed8bc67c57ee3d3c668d00ec2bb96bbef45d)