[dv,pwm] Synchronize with monitor pulse cycles
The monitor and scoreboard are unable to form exact predictions at
the point of channel configuration and enabling, leading to its
present 'SettleTime' behavior.

However, we can spot the first transition in the output, at the end
of the '_X+1' pulses in blinking modes, and thus synchronize exactly
with the monitor/DUT signaling, allowing no further deviation until
the configuration is changed.

Move reporting of configuration to UVM_MEDIUM level so that it's
visible without generating huge logs.

Signed-off-by: Adrian Lees <a.lees@lowrisc.org>