[deps] Manually update Cargo toml/lock to resolve `cargo audit` warnings
This commit includes a variety of updates to the `Cargo.toml` and
`Cargo.lock` files used by OpenTitan to manage rust dependencies, such
that the majority of the issues/vulnerabilities reported by `cargo
audit` are appropriately addressed. This was chosen in lieu of a full
`cargo update` for now, because that seemed to cause some builds to
break.

The packages updated in this way include:
 - `zerocopy` (vulnerable, version yanked).
 - `indicatif` (used `instant`, which is un-maintained).
 - `rsa` (side-channel vulnerability still remains)
 - `openssl` (vulnerability)
 - `mio` (vulnerability)
 - `idna` (vulnerability)
 - `url` (depends on `idna`)
 - `smallvec` (dependency of `url` and `rsa`)

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>