[topgen] Make all flash_ctrl parameter mandatory in the top cfg
Currently we omit certain flash_ctrl parameters in the top cfg
and we rely on the Flash class to fill out the rest using default
values. But it does not really make sense to have default values
for flash. Instead, require all parameters to be specified and
just check the types in the class. Also remove some flash_ctrl
parameter hardcoding that really should not be there. Finally,
make sure that the new flash parameters are serialized in the
complete top cfg.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>