[dv,pwm] Testing of regwen functionality
Introduce a test that attempts to modify the PWM configuration
after the outputs have been set operating and the REGWEN bit has
been cleared.

Model the REGWEN functionality in the scoreboard.

Tidy the reset handling within the scoreboard as required to
support this sequence, since clearing the REGWEN shall only be
achievable by a reset of the DUT. Normal shutdown via the
configuration registers is not possible.

Signed-off-by: Adrian Lees <a.lees@lowrisc.org>