[rom_ctrl, dv] Conditional coverage hole in tlul_adapter_sram fifos
A reported coverage hole by the coverage reports for rvalid_o assignment.
It never sees !empty when under_rst and that is not possible. If empty is
true then wptr_wrap_cnt_q and rptr_wrapt_cnt_q are equal in u_fifo_cnt
and when the fifo is under reset, then both pointers gets reset to 0
regardless of the value of Secure parameter. Thus to resolve the hole, we
tweak the assignment of rvalid_o according to the value of Pass.

Signed-off-by: Kinza Qamar <kqzaman@lowrisc.org>