[opentitantool] Rescue erases other slot by default
Add a default true flag `opentitantool rescue firmware
--erase_other_slot true"` This can be used to control whether the
content of the slot not being newly programmed should be left in place,
or erased.  The latter ensures that the newly "rescued" firmware is the
one to boot, and is the default, opposite of the previous behavior.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>

Change-Id: I6c2fc46553d63d86a87fa6dc17bd130e40152bb8
(cherry picked from commit 7baa7540a522999b87e96db29fe061286bb8f47a)