[opentitantool] Rearrange handling of wait after rescue
The `--wait` flag to `opentitantool rescue firmware` can be used to
control whether the new firmware should immediately boot, or whether the
chip should remain in the rescue bootloader after the invocation
terminates.  The latter is useful if more rescue commands are expected,
or if the caller wants to set up some monitoring, before releasing the
newly flashed code to run.

This CL does not change the behavior, but changes how the former
"immadiate boot" case is implemented, in that it is done by an explicit
reset pulse after finishing the flashing.  This makes the logic simpler
in preparation for the next CL, which introduces a second call to
`rescue.update_firmware()`, which would not be possible unless
`rescue.wait()` has been previously called.

Change-Id: I37b9947d3c70791226eb47041bd70b7d7b31d9d4
Signed-off-by: Jes B. Klinke <jbk@chromium.org>
(cherry picked from commit 9aefafbcd14b324ca492fc4d137bd7e5546d6eb8)