Catch URL error if remote stats bucket does not exist
When creating a new IP or moving an IP to ipgen, there is no initial
regression run. Consequently, fetching its stats fails as the remote
path does not exist.

To unblock this error, this PR catches the URL error if the path does
not exist and returns that 0 runs out of 0 are performed.

Signed-off-by: Robert Schilling <rschilling@rivosinc.com>