[topgen] Introduce a GroupProxy class for clocks
The clock groups are defined in the 'clocks' subsection of the
configuration and used in other places (such as alerts). However
this is suboptional for serialization because the users end up
serializing the whole group which is already defined somewhere else.
This commit introduces a new class GroupProxy that forwards a clock
group but serializes only to the group name.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>